+++ src/stories/chat.js
@@ -69,4 +69,31 @@ storiesOf('Chat', module)
       <Chat historicMessages={historicMessages} messages={messagesForConversation} turnOffLoop styles={styles} />
     </div>
   );
+})
+.add('Chat configured styles and ends at the last message and allows the user to add and scroll messages', () => {
+  const messagesForConversation = [
+    {
+      type: 'typing',
+      duration: 500,
+      inbound: true,
+    },
+    {
+      message: 'Press enter to write a message',
+      inbound: true,
+      backColor: 'white',
+      duration: 1000,
+    },
+  ];
+  const historicMessages = [
+    {
+      message: 'Type in the input box below',
+      inbound: true,
+      backColor: 'white',
+    },
+  ];
+  return (
+    <div>
+      <Chat historicMessages={historicMessages} messages={messagesForConversation} turnOffLoop isScrollable />
+    </div>
+  );
 });